refactor: #43 make utils, test folder in adoorback folder
