+++ backend/adoorback/adoorback/test/mock_db
@@ -5,7 +5,7 @@ from feed.models import Question, Article, Response, Post
 from comment.models import Comment
 from like.models import Like
 from django.contrib.contenttypes.models import ContentType
-from adoorback.utils.mock_seed import set_mock_seed
+from adoorback.test.mock_seed import set_mock_seed
 User = get_user_model()
 
 set_mock_seed()
