+++ backend/adoorback/adoorback/test/seed.py
@@ -6,7 +6,7 @@ from django.utils import timezone
 from django.contrib.auth import get_user_model
 from faker import Faker
 
-from adoorback.content_types import get_comment_type
+from adoorback.utils.content_types import get_comment_type
 from account.models import FriendRequest
 from feed.models import Article, Response, Question, ResponseRequest
 from comment.models import Comment
