+++ backend/adoorback/adoorback/test/shell
@@ -5,8 +5,8 @@ from feed.models import Question, Article, Response, Post, ResponseRequest
 from comment.models import Comment
 from like.models import Like
 from notification.models import Notification
-from adoorback.utils.seed import set_seed
-from adoorback.content_types import get_article_type, get_question_type, get_comment_type
+from adoorback.test.seed import set_seed
+from adoorback.utils.content_types import get_article_type, get_question_type, get_comment_type
 from django.contrib.auth import get_user_model
 User = get_user_model()
 User.objects.values_list('email', flat=True)
