+++ backend/adoorback/like/models.py
@@ -6,11 +6,11 @@ from django.db.models.signals import post_save
 from django.dispatch import receiver
 from django.contrib.auth import get_user_model
 
-from adoorback.content_types import get_comment_type
+from adoorback.utils.content_types import get_comment_type
 from adoorback.models import AdoorTimestampedModel
 from notification.models import Notification
 
-from utils import wrap_content
+from adoorback.utils.helpers import wrap_content
 
 User = get_user_model()
