+++ backend/adoorback/user_report/tests.py
@@ -10,7 +10,7 @@ from feed.models import Article, Question, Response, ResponseRequest
 from comment.models import Comment
 from like.models import Like
 from notification.models import Notification
-from adoorback.utils.seed import set_seed
+from adoorback.test.seed import set_seed
 
 User = get_user_model()
 N = 10
