+++ backend/adoorback/user_report/views.py
@@ -5,9 +5,9 @@ from rest_framework.permissions import IsAuthenticated
 from user_report.models import UserReport
 from user_report.serializers import UserReportSerializer
 
-from adoorback.permissions import IsNotBlocked
-from adoorback.content_types import get_generic_relation_type
-from adoorback.validators import adoor_exception_handler
+from adoorback.utils.permissions import IsNotBlocked
+from adoorback.utils.content_types import get_generic_relation_type
+from adoorback.utils.validators import adoor_exception_handler
 
 
 class UserReportList(generics.CreateAPIView):
