--- backend/adoorback/utils/__init__.py
@@ -1,1 +1,0 @@
-from .wrap_content import *
